Updating capstone.py
Since this code encrypts data when added into the file, I think a different approach is needed. Since I want to see my data, I should not be encrypting it when storing it away. I need to be focused on securing it by finding a way to access the data a certain way.